キャスト式

+++ sample.c
@@ -8,7 +8,7 @@ int main(void)
   printf("整数a："); scanf("%d",&a);
   printf("整数b："); scanf("%d",&b);
 
-  printf("それらの平均は%fです。\n", (a + b) / 2.0);
+  printf("それらの平均は%fです。\n", (double)(a + b) / 2);
 
   return 0;
 }
